feat: moved the last of the old template components
