+++ src/layout/index.js
@@ -1,2 +1,3 @@
 export { default as Container } from './container';
 export { default as Grid } from './grid';
+export { default as Page } from './page';
