+++ src/layout/page/pageDetailsWindow.jsx
