+++ src/surfaces/drawer/drawerDetailsWindow.jsx
